+++ packages/lib/src/components/charts/lume-sparkline-chart/defaults.ts
@@ -1,4 +1,4 @@
-import { LineChartOptions } from '@/composables/options';
+import type { LineChartOptions } from '@/types/options';
 
 export const options: LineChartOptions = {
   margins: {
