+++ packages/lib/src/components/core/lume-axis/composables/y-band-scale-axis.ts
@@ -1,7 +1,7 @@
 import { computed, Ref } from 'vue';
 import { ScaleBand } from 'd3';
 
-import { AxisOptions } from '@/composables/options';
+import type { AxisOptions } from '@/types/options';
 import { AXIS_GHOST_PADDING } from '@/utils/constants';
 import { AxisMixin } from '../types';
