+++ packages/lib/src/components/core/lume-axis/index.ts
@@ -1,1 +1,2 @@
 export { default } from './lume-axis.vue';
+export * from './types';
