+++ packages/lib/src/composables/group-props.ts
@@ -1,12 +1,12 @@
 import { PropType } from 'vue';
 
-import { Options } from '@/composables/options';
 import { orientationValidator } from '@/composables/props';
 import { Scale } from '@/composables/scales';
 
 import { Orientation, ORIENTATIONS } from '@/utils/constants';
-import { DatasetValueObject, InternalData } from '@/types/dataset';
-import { ContainerSize } from '@/types/size';
+import type { DatasetValueObject, InternalData } from '@/types/dataset';
+import type { ContainerSize } from '@/types/size';
+import type { Options } from '@/types/options';
 
 export const withGroupProps = <
   T extends Options = Options,
