+++ packages/lib/src/composables/negative-values.ts
@@ -3,9 +3,10 @@ import { computed, Ref } from 'vue';
 import { flatValues } from '@/utils/helpers';
 
 import { Orientation, ORIENTATIONS } from '@/utils/constants';
-import { InternalData } from '@/types/dataset';
-import { ContainerSize } from '@/types/size';
 import { Scale } from './scales';
+
+import type { InternalData } from '@/types/dataset';
+import type { ContainerSize } from '@/types/size';
 
 export function checkNegativeValues(data: Ref<InternalData>) {
   const hasNegativeValues = computed(() =>
