+++ packages/lib/src/playground/custom-tooltip.vue
@@ -13,7 +13,7 @@ import { defineComponent, PropType, ref, toRefs, watch } from 'vue';
 
 import { useBase } from '@/composables/base';
 
-import { Data } from '@/types/dataset';
+import type { Data } from '@/types/dataset';
 
 const EMOJI_MAP = {
   positive: '🚀',
