+++ packages/lib/src/types/index.ts
@@ -1,0 +1,5 @@
+export * from './alluvial';
+export * from './dataset';
+export * from './events';
+export * from './options';
+export * from './size';
